sini: split object into two - section and key arguments
Previously sini was using only <object> argument in format
<section>.<key>, this was preventing use of dots in <section>,
since sini was considering such dot as a separator.

By spliting that argument into two arguments this problem was
fixed. Also it make is more clean to retrieve keys that are not
in any section since <object> can now be ommited.

Signed-off-by: Michał Łyszczek <michal.lyszczek@bofc.pl>